+++ NEWS.md
@@ -1,3 +1,12 @@
+Version 1.3.0: October 5, 2023
+-------------------------------------------------------------------------------
+
+General:
+
++ Updated dependencies alakazam >= 1.3.0, shazam >= 1.2.0, ggplot2 >= 3.4.0.
++ Added a `locus` column to the package's example data, to satisfy 
+  alakazam >= 1.3.0 requirements.
+
 Version 1.2.1: September 21, 2022
 -------------------------------------------------------------------------------